move test/emoji.js into test/emoji.js
